+++ test/jsx.test.tsx
@@ -1,0 +1,10 @@
+/* eslint-disable unused-imports/no-unused-vars */
+
+// 允许同一行有多个 jsx 元素
+const b = (
+  <div>
+    <div>Hello</div>
+    当前{ true ? '已' : '未' }登录
+    开发工具 { 'Mixte' + ' ' }
+  </div>
+);